chore: refactor backend structure
