+++ backend/go.mod
@@ -1,4 +1,4 @@
-module techinternships-backend
+module backend
 
 go 1.21.0
