+++ backend/models/models.go
@@ -1,4 +1,4 @@
-package main
+package models
 
 type Internship struct {
 	Company   string `json:"company"`
